requirements.txt
alembic==1.7.7
autopep8==1.6.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Dance==5.1.0
Flask-Limiter==2.2.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
limits==2.4.0
Mako==1.2.0
MarkupSafe==2.1.1
oauthlib==3.2.0
pyclean==2.1.0
pycodestyle==2.8.0
pycparser==2.21
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.32
toml==0.10.2
urllib3==1.26.9
URLObject==2.4.3
Werkzeug==2.0.3
WTForms==3.0.1
